requirements/dev.txt
black
flake8
flake8-black
isort>=5.7
pytest
pytest-cov
ptflops>=0.6.8
numpy
pytorch_lightning
onnxruntime~=1.13
